refactor: rewrite Overlay of PurePanel (#50097)
