+++ components/popover/PurePanel.tsx
@@ -7,18 +7,20 @@ import { getRenderPropValue } from '../_util/getRenderPropValue';
 import { ConfigContext } from '../config-provider';
 import useStyle from './style';
 
-export const getOverlay = (
-  prefixCls?: string,
-  title?: PopoverProps['title'],
-  content?: PopoverProps['content'],
-) => {
+interface OverlayProps {
+  prefixCls?: string;
+  title?: React.ReactNode;
+  content?: React.ReactNode;
+}
+
+export const Overlay: React.FC<OverlayProps> = ({ title, content, prefixCls }) => {
   if (!title && !content) {
     return null;
   }
   return (
     <>
-      {title && <div className={`${prefixCls}-title`}>{getRenderPropValue(title)}</div>}
-      <div className={`${prefixCls}-inner-content`}>{getRenderPropValue(content)}</div>
+      {title && <div className={`${prefixCls}-title`}>{title}</div>}
+      {content && <div className={`${prefixCls}-inner-content`}>{content}</div>}
     </>
   );
 };
@@ -43,20 +45,22 @@ export const RawPurePanel: React.FC<RawPurePanelProps> = (props) => {
     children,
   } = props;
 
+  const titleNode = getRenderPropValue(title);
+  const contentNode = getRenderPropValue(content);
+
+  const cls = classNames(
+    hashId,
+    prefixCls,
+    `${prefixCls}-pure`,
+    `${prefixCls}-placement-${placement}`,
+    className,
+  );
+
   return (
-    <div
-      className={classNames(
-        hashId,
-        prefixCls,
-        `${prefixCls}-pure`,
-        `${prefixCls}-placement-${placement}`,
-        className,
-      )}
-      style={style}
-    >
+    <div className={cls} style={style}>
       <div className={`${prefixCls}-arrow`} />
       <Popup {...props} className={hashId} prefixCls={prefixCls}>
-        {children || getOverlay(prefixCls, title, content)}
+        {children || <Overlay prefixCls={prefixCls} title={titleNode} content={contentNode} />}
       </Popup>
     </div>
   );
